Fix how Intake handles multiValued entries. Initial patch to IntegerField provided by Derek Gaasch.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/branches/TURBINE_2_3_BRANCH@221831 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/turbine/services/intake/model/DoubleField.java
@@ -113,18 +113,21 @@ public class DoubleField
         if (isMultiValued)
         {
             String[] inputs = parser.getStrings(getKey());
-            Double[] values = new Double[inputs.length];
+            double[] values = new double[inputs.length];
             for (int i = 0; i < inputs.length; i++)
             {
                 values[i] = StringUtils.isNotEmpty(inputs[i])
-                        ? new Double(inputs[i]) : (Double) getEmptyValue();
+                        ? new Double(inputs[i]).doubleValue() 
+                        : ((Double) getEmptyValue()).doubleValue();
             }
             setTestValue(values);
         }
         else
         {
             String val = parser.getString(getKey());
-            setTestValue(StringUtils.isNotEmpty(val) ? new Double(val) : (Double) getEmptyValue());
+            setTestValue(StringUtils.isNotEmpty(val) 
+                    ? new Double(val) : (Double) getEmptyValue());
         }
     }
+    
 }
